projects/hermes-memory-system/requirements.txt
PyYAML>=6.0
sqlite3-python
